refactor(e2e): helper functions to commands
